+++ JosueZelaya_Lab3/src/josuezelaya_lab3/Tramontana.java
@@ -13,36 +13,36 @@ import java.util.Date;
  * @author usuario
  */
 public class Tramontana extends Carro {
-    private int peso;
-    private int transmision;
+    private double peso;
+    private double transmision;
 
     public Tramontana() {
         super();
     }
 
    
-    public Tramontana(int peso, int transmision, int serie, Date fecha, Color color, String marca_llantas, String polarizado, int velocidad_max, int km_galon, int precio_venta) {
+    public Tramontana(double peso, double transmision, int serie, Date fecha, Color color, String marca_llantas, String polarizado, double velocidad_max, double km_galon, double precio_venta) {
         super(serie, fecha, color, marca_llantas, polarizado, velocidad_max, km_galon, precio_venta);
         this.peso = peso;
         this.transmision= transmision;
     }
 
-    public int getPeso() {
+    public double getPeso() {
         return peso;
     }
 
-    public void setPeso(int peso) {
+    public void setPeso(double peso) {
         if (peso>=1276 && peso<=1376){
             this.peso = peso;
         }
         
     }
 
-    public int getTransmision() {
+    public double getTransmision() {
         return transmision;
     }
 
-    public void setTransmision(int transmision) {
+    public void setTransmision(double transmision) {
         if (transmision>=6 && transmision<=7){
             this.transmision = transmision;
         }
